+++ app/src/test/kotlin/xyz/mpv/rex/utils/media/StreamStatsFetcherTest.kt
@@ -21,4 +21,11 @@ class StreamStatsFetcherTest {
     assertEquals("0 KB/s", StreamStatsFetcher.formatSpeed(0))
     assertEquals("1.3 MB/s", StreamStatsFetcher.formatSpeed(1_300_000))
   }
+
+  @Test
+  fun parsesPingLatencyAndFailure() {
+    assertEquals(50, StreamStatsFetcher.parsePingMs("64 bytes from host: time=50.7 ms"))
+    assertEquals(1, StreamStatsFetcher.parsePingMs("64 bytes from host: time<1 ms"))
+    assertEquals(0, StreamStatsFetcher.parsePingMs("100% packet loss"))
+  }
 }